+++ NEWS.md
@@ -1,3 +1,10 @@
+# armgin 0.1.0.9000
+
+* Summarise by individual grouping variables one at a time with `method
+    = "individual"`.  The argument `hierarchy` is deprecated and replaced by
+    `method`, which can be one of  `"hierarchy"`, `"combination"` or
+    `"individual"`.
+
 # armgin 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.